requirements.txt
aniso8601==9.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
distlib==0.3.3
filelock==3.4.0
Flask==2.0.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
marshmallow==3.14.1
pipenv==2021.11.23
platformdirs==2.4.0
PyJWT==2.3.0
pyrsistent==0.18.0
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
zipp==3.6.0
flasgger==0.9.5